+++ WEB-INF/src/com/krishagni/catissueplus/core/biospecimen/print/SpecimenQtyPrintToken.java
@@ -1,5 +1,7 @@
 package com.krishagni.catissueplus.core.biospecimen.print;
 
+import java.math.BigDecimal;
+
 import com.krishagni.catissueplus.core.biospecimen.domain.Specimen;
 import com.krishagni.catissueplus.core.common.domain.AbstractLabelTmplToken;
 import com.krishagni.catissueplus.core.common.domain.LabelTmplToken;
@@ -13,8 +15,8 @@ public class SpecimenQtyPrintToken extends AbstractLabelTmplToken implements Lab
 
 	@Override
 	public String getReplacement(Object object) {
-		Specimen specimen = (Specimen)object;
-		Double dbl = specimen.getInitialQuantity();
+		Specimen specimen = (Specimen) object;
+		BigDecimal dbl = specimen.getInitialQuantity();
 		return dbl != null ? dbl.toString() : null;
 	}
 }
